[skip ci] wip, working suite-dev docker-compose recipe
