+++ docker/suite-dev/Dockerfile
@@ -1,6 +1,9 @@
 FROM node:12-buster
 
 RUN apt-get update
-# todo: add libs caching based on libs checksum
+RUN apt install rsync grsync -y
 
+RUN mkdir suite-copy
 WORKDIR /suite
+
+COPY ./suite-dev/run-inside.sh /
