--- docker/suite-test.docker-compose.yml
@@ -1,31 +1,0 @@
-version: '3'
-services:
-  trezor-env:
-    build:
-      context: . # . stands for pwd, so docker-compose must be called from root
-      dockerfile: ./trezor-env/Dockerfile
-    network_mode: "host"
-    environment:
-      - DISPLAY=unix$DISPLAY
-    volumes:
-      - /tmp/.X11-unix:/tmp/.X11-unix
-    ports:
-      - "9001:9001"
-    # command: ["./python/bin/trezor.elf"]
-    command: ["./python/bin/trezor-emu-core-v2.2.0", "-O0", "-X", "heapsize=20M", "-m", "main"]
-    # command: >
-    #   bash -c "./python/bin/trezor-emu-core-v2.2.0 -O0 -X heapsize=20M -m main"
-    # command: >
-    #   bash -c "python3 ./python/main.py"
-        
-  # bridge:
-  #   build:
-  #     context: .
-  #     dockerfile: ./packages/integration-tests/docker/bridge/Dockerfile
-  #   #this works by some dark magic
-  #   ports:
-  #     - "127.0.0.1:21325:11325"
-  #   network_mode: "host"
-  #   command: >
-  #     bash -c "redir --lport 11325 --cport 21325 --caddr 127.0.0.1 &
-  #     ./trezord-go -e 21324 -u=false"
